pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "siemens-mri-video-analyzer"
version = "0.1.0"
description = "Extract numerical MRI console data from Siemens MPS ramp videos"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.10",
    "pyyaml>=6.0",
    "click>=8.1.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
]

[project.scripts]
smva = "smva.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

